extra casting for other returns lof lexsort in tensorflow

+++ ivy/functional/backends/tensorflow/experimental/sorting.py
@@ -23,14 +23,14 @@ def lexsort(
 ) -> Union[tf.Tensor, tf.Variable]:
     shape = keys.shape.as_list()
     if len(shape) == 1:
-        return tf.argsort(keys, axis=axis, stable=True)
+        return tf.cast(tf.argsort(keys, axis=axis, stable=True), tf.int64)
     if shape[0] == 0:
         raise TypeError('need sequence of keys with len > 0 in lexsort')
     if len(shape) == 2 and shape[1] == 1:
-        return tf.convert_to_tensor([0])
+        return tf.cast(tf.convert_to_tensor([0]), tf.int64)
     result = tf.argsort(keys[0], axis=axis, stable=True)
     if shape[0] == 1:
-        return result
+        return tf.cast(result, tf.int64)
     for i in range(1, shape[0]):
         key = keys[i]
         ind = tf.gather(key, result)
